requirements.txt
Flask==3.1.1
flask-cors
python-dotenv==1.1.1
Requests==2.32.4
